Harden Cotsel maintainability seams
