+++ gateway/src/core/authSessionClient.ts
@@ -55,8 +55,80 @@ export interface AuthSessionClient {
   checkReadiness(requestId?: string): Promise<void>;
 }
 
+const AUTH_SERVICE_ROLES: AuthServiceRole[] = ['buyer', 'supplier', 'admin', 'oracle'];
+const OPERATOR_CAPABILITIES: OperatorCapability[] = [
+  'governance:write',
+  'compliance:write',
+  'treasury:read',
+  'treasury:prepare',
+  'treasury:approve',
+  'treasury:execute_match',
+  'treasury:close',
+];
+const SIGNER_ACTION_CLASSES: SignerActionClass[] = [
+  'governance',
+  'treasury_approve',
+  'treasury_execute',
+  'treasury_close',
+  'compliance_sensitive',
+  'emergency_admin',
+];
+
 function buildUrl(baseUrl: string, pathname: string): string {
   return `${baseUrl}${pathname}`;
+}
+
+function isStringOrNull(value: unknown): value is string | null {
+  return typeof value === 'string' || value === null;
+}
+
+function isOptionalStringOrNull(value: unknown): value is string | null | undefined {
+  return value === undefined || isStringOrNull(value);
+}
+
+function isKnownValue<T extends string>(allowedValues: readonly T[], value: unknown): value is T {
+  return typeof value === 'string' && allowedValues.includes(value as T);
+}
+
+function isSignerAuthorization(value: unknown): value is SignerAuthorization {
+  if (!value || typeof value !== 'object' || Array.isArray(value)) {
+    return false;
+  }
+
+  const candidate = value as Record<string, unknown>;
+  return (
+    typeof candidate.bindingId === 'string' &&
+    typeof candidate.walletAddress === 'string' &&
+    isKnownValue(SIGNER_ACTION_CLASSES, candidate.actionClass) &&
+    typeof candidate.environment === 'string' &&
+    typeof candidate.approvedAt === 'string' &&
+    typeof candidate.approvedBy === 'string' &&
+    isStringOrNull(candidate.ticketRef) &&
+    isStringOrNull(candidate.notes)
+  );
+}
+
+function isAuthSession(value: unknown): value is AuthSession {
+  if (!value || typeof value !== 'object' || Array.isArray(value)) {
+    return false;
+  }
+
+  const candidate = value as Record<string, unknown>;
+  return (
+    typeof candidate.userId === 'string' &&
+    isOptionalStringOrNull(candidate.accountId) &&
+    isStringOrNull(candidate.walletAddress) &&
+    isKnownValue(AUTH_SERVICE_ROLES, candidate.role) &&
+    Array.isArray(candidate.capabilities) &&
+    candidate.capabilities.every((capability) => isKnownValue(OPERATOR_CAPABILITIES, capability)) &&
+    Array.isArray(candidate.signerAuthorizations) &&
+    candidate.signerAuthorizations.every(isSignerAuthorization) &&
+    typeof candidate.issuedAt === 'number' &&
+    Number.isFinite(candidate.issuedAt) &&
+    typeof candidate.expiresAt === 'number' &&
+    Number.isFinite(candidate.expiresAt) &&
+    isOptionalStringOrNull(candidate.email)
+  );
 }
 
 export function createAuthSessionClient(config: GatewayConfig): AuthSessionClient {
@@ -87,8 +159,8 @@ export function createAuthSessionClient(config: GatewayConfig): AuthSessionClien
         });
       }
 
-      const payload = (await response.json()) as SessionResponse;
-      if (!payload.success || !payload.data) {
+      const payload = (await response.json().catch(() => null)) as SessionResponse | null;
+      if (!payload?.success || !isAuthSession(payload.data)) {
         throw new GatewayError(
           503,
           'UPSTREAM_UNAVAILABLE',
