+++ gateway/src/routes/governanceMutations.ts
@@ -1,7 +1,7 @@
 /**
  * SPDX-License-Identifier: Apache-2.0
  */
-import { NextFunction, Request, Response, Router } from 'express';
+import { NextFunction, Response, Router } from 'express';
 import { GatewayConfig } from '../config/env';
 import { AuthSessionClient } from '../core/authSessionClient';
 import {
@@ -11,19 +11,12 @@ import {
   validateGovernanceAuditInput,
   validateProposalId,
 } from '../core/governanceMutationService';
-import type {
-  GovernanceActionPrepared,
-  GovernanceBroadcastConfirmed,
-} from '../core/governanceMutationService';
-import type { SignerAuthorization } from '../core/authSessionClient';
 import { GovernanceMutationPreflightReader } from '../core/governanceStatusService';
 import { IdempotencyStore } from '../core/idempotencyStore';
 import {
   createAuthenticationMiddleware,
-  requireAuthorizedSignerBinding,
   requireMutationWriteAccess,
   requireOperatorActionCapability,
-  requireSignerWalletAddress,
   requireWalletBoundSession,
   resolveGatewayActorKey,
 } from '../middleware/auth';
@@ -36,6 +29,14 @@ import {
   GatewayErrorHandlerWorkflow,
   type GovernanceReplaySpec,
 } from '../core/errorHandlerWorkflow';
+import { registerGovernanceDirectSignRoutes } from './governanceDirectSignMutations';
+import {
+  getMutationContext,
+  getPathParam,
+  type MutationRequest,
+  nowSeconds,
+  requireDirectSignGovernancePath,
+} from './governanceMutationRouteSupport';
 
 export interface GovernanceMutationRouterOptions {
   authSessionClient: AuthSessionClient;
@@ -44,138 +45,6 @@ export interface GovernanceMutationRouterOptions {
   mutationService: GovernanceMutationService;
   idempotencyStore: IdempotencyStore;
   failedOperationWorkflow?: GatewayErrorHandlerWorkflow;
-}
-
-interface MutationContext {
-  principal: GatewayPrincipal;
-  requestContext: RequestContext;
-  idempotencyKey: string;
-}
-
-type MutationRequest = Request<Record<string, string | string[]>, unknown, Record<string, unknown>>;
-
-function nowSeconds(): number {
-  return Math.floor(Date.now() / 1000);
-}
-
-function getMutationContext(req: MutationRequest): MutationContext {
-  if (!req.gatewayPrincipal) {
-    throw new GatewayError(401, 'AUTH_REQUIRED', 'Authentication is required');
-  }
-
-  if (!req.requestContext) {
-    throw new GatewayError(500, 'INTERNAL_ERROR', 'Request context was not initialized');
-  }
-
-  if (!req.idempotencyState?.idempotencyKey) {
-    throw new GatewayError(500, 'INTERNAL_ERROR', 'Idempotency context was not initialized');
-  }
-
-  return {
-    principal: req.gatewayPrincipal,
-    requestContext: req.requestContext,
-    idempotencyKey: req.idempotencyState.idempotencyKey,
-  };
-}
-
-function getAuthorizedGovernanceSigner(
-  req: MutationRequest,
-  config: GatewayConfig,
-  actionDescription: string,
-): { signerWallet: string; signerBinding: SignerAuthorization } {
-  if (!req.gatewayPrincipal) {
-    throw new GatewayError(401, 'AUTH_REQUIRED', 'Authentication is required');
-  }
-
-  const signerWallet = requireSignerWalletAddress(
-    typeof req.body?.signerWallet === 'string' ? req.body.signerWallet : null,
-  );
-  const signerBinding = requireAuthorizedSignerBinding(
-    req.gatewayPrincipal,
-    config,
-    'governance',
-    signerWallet,
-    actionDescription,
-  );
-
-  return { signerWallet, signerBinding };
-}
-
-function isAllowedHumanGovernanceMutationPath(path: string): boolean {
-  return path.endsWith('/prepare') || /^\/actions\/[^/]+\/confirm$/.test(path);
-}
-
-function requireDirectSignGovernancePath(req: Request, _res: Response, next: NextFunction): void {
-  if (req.method === 'POST' && !isAllowedHumanGovernanceMutationPath(req.path)) {
-    next(
-      new GatewayError(
-        409,
-        'CONFLICT',
-        'Human governance queue routes are retired; use the corresponding /prepare endpoint and then /governance/actions/:actionId/confirm',
-        {
-          reason: 'direct_sign_required',
-          route: req.path,
-        },
-      ),
-    );
-    return;
-  }
-
-  next();
-}
-
-function getPathParam(value: string | string[] | undefined, field: string): string | undefined {
-  if (Array.isArray(value)) {
-    throw new GatewayError(400, 'VALIDATION_ERROR', `Path parameter ${field} must be a string`);
-  }
-
-  return value;
-}
-
-async function prepareAndRespond(
-  req: MutationRequest,
-  res: Response,
-  next: NextFunction,
-  config: GatewayConfig,
-  actionFactory: (
-    signerWallet: string,
-    signerBinding: SignerAuthorization,
-  ) => Promise<GovernanceActionPrepared>,
-): Promise<void> {
-  try {
-    const { signerWallet, signerBinding } = getAuthorizedGovernanceSigner(
-      req,
-      config,
-      'Preparing privileged governance approval',
-    );
-    const prepared = await actionFactory(signerWallet, signerBinding);
-    res.status(200).json(successResponse(prepared));
-  } catch (error) {
-    next(error);
-  }
-}
-
-async function confirmAndRespond(
-  req: MutationRequest,
-  res: Response,
-  next: NextFunction,
-  config: GatewayConfig,
-  actionFactory: (
-    signerWallet: string,
-    signerBinding: SignerAuthorization,
-  ) => Promise<GovernanceBroadcastConfirmed>,
-): Promise<void> {
-  try {
-    const { signerWallet, signerBinding } = getAuthorizedGovernanceSigner(
-      req,
-      config,
-      'Confirming privileged governance broadcast',
-    );
-    const confirmed = await actionFactory(signerWallet, signerBinding);
-    res.status(200).json(successResponse(confirmed));
-  } catch (error) {
-    next(error);
-  }
 }
 
 async function queueAndRespond(
@@ -1153,626 +1022,7 @@ export function createGovernanceMutationRouter(options: GovernanceMutationRouter
       ),
   );
 
-  // ── Direct-sign prepare endpoints ────────────────────────────────────────
-  // Each endpoint validates pre-flight state, builds the canonical txRequest,
-  // records audit intent, and returns the payload for the admin wallet to sign.
-  // The admin signs and broadcasts independently, then calls /confirm.
-
-  router.post('/governance/pause/prepare', idempotency, (req, res, next) =>
-    prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-      const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-      const audit = validateGovernanceAuditInput(req.body);
-      const status = await options.governanceReader.getGovernanceStatus();
-      if (status.paused) {
-        throw new GatewayError(409, 'CONFLICT', 'Protocol is already paused');
-      }
-
-      return options.mutationService.prepareAction({
-        category: 'pause',
-        contractMethod: 'pause',
-        routePath: req.originalUrl || req.path,
-        audit,
-        principal,
-        signerWallet,
-        requestContext,
-        signerBinding,
-        idempotencyKey,
-      });
-    }),
-  );
-
-  router.post('/governance/unpause/proposal/prepare', idempotency, (req, res, next) =>
-    prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-      const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-      const audit = validateGovernanceAuditInput(req.body);
-      const status = await options.governanceReader.getGovernanceStatus();
-      if (!status.paused) {
-        throw new GatewayError(
-          409,
-          'CONFLICT',
-          'Protocol must be paused before creating an unpause proposal',
-        );
-      }
-      if (!status.oracleActive) {
-        throw new GatewayError(
-          409,
-          'CONFLICT',
-          'Oracle must be active before creating an unpause proposal',
-        );
-      }
-
-      return options.mutationService.prepareAction({
-        category: 'unpause',
-        contractMethod: 'proposeUnpause',
-        routePath: req.originalUrl || req.path,
-        audit,
-        principal,
-        signerWallet,
-        requestContext,
-        signerBinding,
-        idempotencyKey,
-      });
-    }),
-  );
-
-  router.post('/governance/unpause/proposal/approve/prepare', idempotency, (req, res, next) =>
-    prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-      const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-      const audit = validateGovernanceAuditInput(req.body);
-      const proposal = await options.governanceReader.getUnpauseProposalState();
-      if (!proposal.hasActiveProposal) {
-        throw new GatewayError(
-          409,
-          'CONFLICT',
-          'No active unpause proposal is available to approve',
-        );
-      }
-
-      if (await options.governanceReader.hasApprovedUnpause(signerWallet)) {
-        throw new GatewayError(
-          409,
-          'CONFLICT',
-          'Caller has already approved the active unpause proposal',
-        );
-      }
-
-      return options.mutationService.prepareAction({
-        category: 'unpause',
-        contractMethod: 'approveUnpause',
-        routePath: req.originalUrl || req.path,
-        audit,
-        principal,
-        signerWallet,
-        requestContext,
-        signerBinding,
-        idempotencyKey,
-      });
-    }),
-  );
-
-  router.post('/governance/unpause/proposal/cancel/prepare', idempotency, (req, res, next) =>
-    prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-      const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-      const audit = validateGovernanceAuditInput(req.body);
-      const proposal = await options.governanceReader.getUnpauseProposalState();
-      if (!proposal.hasActiveProposal) {
-        throw new GatewayError(
-          409,
-          'CONFLICT',
-          'No active unpause proposal is available to cancel',
-        );
-      }
-
-      return options.mutationService.prepareAction({
-        category: 'unpause',
-        contractMethod: 'cancelUnpauseProposal',
-        routePath: req.originalUrl || req.path,
-        audit,
-        principal,
-        signerWallet,
-        requestContext,
-        signerBinding,
-        idempotencyKey,
-      });
-    }),
-  );
-
-  router.post('/governance/claims/pause/prepare', idempotency, (req, res, next) =>
-    prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-      const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-      const audit = validateGovernanceAuditInput(req.body);
-      const status = await options.governanceReader.getGovernanceStatus();
-      if (status.claimsPaused) {
-        throw new GatewayError(409, 'CONFLICT', 'Claims are already paused');
-      }
-
-      return options.mutationService.prepareAction({
-        category: 'claims_pause',
-        contractMethod: 'pauseClaims',
-        routePath: req.originalUrl || req.path,
-        audit,
-        principal,
-        signerWallet,
-        requestContext,
-        signerBinding,
-        idempotencyKey,
-      });
-    }),
-  );
-
-  router.post('/governance/claims/unpause/prepare', idempotency, (req, res, next) =>
-    prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-      const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-      const audit = validateGovernanceAuditInput(req.body);
-      const status = await options.governanceReader.getGovernanceStatus();
-      if (!status.claimsPaused) {
-        throw new GatewayError(409, 'CONFLICT', 'Claims are not currently paused');
-      }
-
-      return options.mutationService.prepareAction({
-        category: 'claims_unpause',
-        contractMethod: 'unpauseClaims',
-        routePath: req.originalUrl || req.path,
-        audit,
-        principal,
-        signerWallet,
-        requestContext,
-        signerBinding,
-        idempotencyKey,
-      });
-    }),
-  );
-
-  router.post('/governance/treasury/sweep/prepare', idempotency, (req, res, next) =>
-    prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-      const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-      const audit = validateGovernanceAuditInput(req.body);
-      const status = await options.governanceReader.getGovernanceStatus();
-      if (status.claimsPaused) {
-        throw new GatewayError(
-          409,
-          'CONFLICT',
-          'Treasury sweep is unavailable while claims are paused',
-        );
-      }
-
-      const claimableBalance = await options.governanceReader.getTreasuryClaimableBalance();
-      if (claimableBalance <= 0n) {
-        throw new GatewayError(409, 'CONFLICT', 'Treasury claimable balance is zero');
-      }
-
-      return options.mutationService.prepareAction({
-        category: 'treasury_sweep',
-        contractMethod: 'claimTreasury',
-        routePath: req.originalUrl || req.path,
-        audit,
-        principal,
-        signerWallet,
-        requestContext,
-        signerBinding,
-        idempotencyKey,
-      });
-    }),
-  );
-
-  router.post(
-    '/governance/treasury/payout-receiver/proposals/prepare',
-    idempotency,
-    (req, res, next) =>
-      prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-        const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-        const audit = validateGovernanceAuditInput(req.body);
-        const status = await options.governanceReader.getGovernanceStatus();
-        const newPayoutReceiver = validateAddressInput(
-          (req.body as Record<string, unknown>)?.newPayoutReceiver,
-          'newPayoutReceiver',
-        );
-        if (newPayoutReceiver.toLowerCase() === status.treasuryPayoutAddress.toLowerCase()) {
-          throw new GatewayError(
-            409,
-            'CONFLICT',
-            'New payout receiver matches the current treasury payout receiver',
-          );
-        }
-
-        return options.mutationService.prepareAction({
-          category: 'treasury_payout_receiver_update',
-          contractMethod: 'proposeTreasuryPayoutAddressUpdate',
-          routePath: req.originalUrl || req.path,
-          audit,
-          principal,
-          signerWallet,
-          requestContext,
-          signerBinding,
-          idempotencyKey,
-          targetAddress: newPayoutReceiver,
-        });
-      }),
-  );
-
-  router.post(
-    '/governance/treasury/payout-receiver/proposals/:proposalId/approve/prepare',
-    idempotency,
-    (req, res, next) =>
-      prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-        const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-        const audit = validateGovernanceAuditInput(req.body);
-        const proposalId = validateProposalId(getPathParam(req.params.proposalId, 'proposalId'));
-        const proposal =
-          await options.governanceReader.getTreasuryPayoutReceiverProposalState(proposalId);
-        if (!proposal) {
-          throw new GatewayError(404, 'NOT_FOUND', 'Treasury payout receiver proposal not found', {
-            proposalId,
-          });
-        }
-        if (proposal.executed || proposal.cancelled || proposal.expired) {
-          throw new GatewayError(
-            409,
-            'CONFLICT',
-            'Treasury payout receiver proposal is no longer approvable',
-            { proposalId },
-          );
-        }
-
-        if (
-          await options.governanceReader.hasApprovedTreasuryPayoutReceiverProposal(
-            proposalId,
-            signerWallet,
-          )
-        ) {
-          throw new GatewayError(
-            409,
-            'CONFLICT',
-            'Caller has already approved this treasury payout receiver proposal',
-            { proposalId },
-          );
-        }
-
-        return options.mutationService.prepareAction({
-          category: 'treasury_payout_receiver_update',
-          contractMethod: 'approveTreasuryPayoutAddressUpdate',
-          routePath: req.originalUrl || req.path,
-          audit,
-          principal,
-          signerWallet,
-          requestContext,
-          signerBinding,
-          idempotencyKey,
-          proposalId,
-          targetAddress: proposal.targetAddress,
-        });
-      }),
-  );
-
-  router.post(
-    '/governance/treasury/payout-receiver/proposals/:proposalId/execute/prepare',
-    idempotency,
-    (req, res, next) =>
-      prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-        const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-        const audit = validateGovernanceAuditInput(req.body);
-        const proposalId = validateProposalId(getPathParam(req.params.proposalId, 'proposalId'));
-        const proposal =
-          await options.governanceReader.getTreasuryPayoutReceiverProposalState(proposalId);
-        if (!proposal) {
-          throw new GatewayError(404, 'NOT_FOUND', 'Treasury payout receiver proposal not found', {
-            proposalId,
-          });
-        }
-        if (proposal.executed || proposal.cancelled || proposal.expired) {
-          throw new GatewayError(
-            409,
-            'CONFLICT',
-            'Treasury payout receiver proposal is not executable',
-            { proposalId },
-          );
-        }
-
-        const status = await options.governanceReader.getGovernanceStatus();
-        if (proposal.approvalCount < status.governanceApprovalsRequired) {
-          throw new GatewayError(
-            409,
-            'CONFLICT',
-            'Treasury payout receiver proposal does not have enough approvals',
-            { proposalId },
-          );
-        }
-        if (proposal.etaSeconds > nowSeconds()) {
-          throw new GatewayError(
-            409,
-            'CONFLICT',
-            'Treasury payout receiver proposal timelock has not elapsed',
-            { proposalId },
-          );
-        }
-
-        return options.mutationService.prepareAction({
-          category: 'treasury_payout_receiver_update',
-          contractMethod: 'executeTreasuryPayoutAddressUpdate',
-          routePath: req.originalUrl || req.path,
-          audit,
-          principal,
-          signerWallet,
-          requestContext,
-          signerBinding,
-          idempotencyKey,
-          proposalId,
-          targetAddress: proposal.targetAddress,
-        });
-      }),
-  );
-
-  router.post(
-    '/governance/treasury/payout-receiver/proposals/:proposalId/cancel-expired/prepare',
-    idempotency,
-    (req, res, next) =>
-      prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-        const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-        const audit = validateGovernanceAuditInput(req.body);
-        const proposalId = validateProposalId(getPathParam(req.params.proposalId, 'proposalId'));
-        const proposal =
-          await options.governanceReader.getTreasuryPayoutReceiverProposalState(proposalId);
-        if (!proposal) {
-          throw new GatewayError(404, 'NOT_FOUND', 'Treasury payout receiver proposal not found', {
-            proposalId,
-          });
-        }
-        if (proposal.executed || proposal.cancelled || !proposal.expired) {
-          throw new GatewayError(
-            409,
-            'CONFLICT',
-            'Treasury payout receiver proposal is not cancellable as expired',
-            { proposalId },
-          );
-        }
-
-        return options.mutationService.prepareAction({
-          category: 'treasury_payout_receiver_update',
-          contractMethod: 'cancelExpiredTreasuryPayoutAddressUpdateProposal',
-          routePath: req.originalUrl || req.path,
-          audit,
-          principal,
-          signerWallet,
-          requestContext,
-          signerBinding,
-          idempotencyKey,
-          proposalId,
-          targetAddress: proposal.targetAddress,
-        });
-      }),
-  );
-
-  router.post('/governance/oracle/disable-emergency/prepare', idempotency, (req, res, next) =>
-    prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-      const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-      const audit = validateGovernanceAuditInput(req.body);
-      const status = await options.governanceReader.getGovernanceStatus();
-      if (!status.oracleActive) {
-        throw new GatewayError(409, 'CONFLICT', 'Oracle is already disabled');
-      }
-
-      return options.mutationService.prepareAction({
-        category: 'oracle_disable_emergency',
-        contractMethod: 'disableOracleEmergency',
-        routePath: req.originalUrl || req.path,
-        audit,
-        principal,
-        signerWallet,
-        requestContext,
-        signerBinding,
-        idempotencyKey,
-        targetAddress: status.oracleAddress,
-      });
-    }),
-  );
-
-  router.post('/governance/oracle/proposals/prepare', idempotency, (req, res, next) =>
-    prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-      const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-      const audit = validateGovernanceAuditInput(req.body);
-      const status = await options.governanceReader.getGovernanceStatus();
-      const newOracleAddress = validateAddressInput(
-        (req.body as Record<string, unknown>)?.newOracleAddress,
-        'newOracleAddress',
-      );
-      if (newOracleAddress.toLowerCase() === status.oracleAddress.toLowerCase()) {
-        throw new GatewayError(
-          409,
-          'CONFLICT',
-          'New oracle address matches the current oracle address',
-        );
-      }
-
-      return options.mutationService.prepareAction({
-        category: 'oracle_update',
-        contractMethod: 'proposeOracleUpdate',
-        routePath: req.originalUrl || req.path,
-        audit,
-        principal,
-        signerWallet,
-        requestContext,
-        signerBinding,
-        idempotencyKey,
-        targetAddress: newOracleAddress,
-      });
-    }),
-  );
-
-  router.post(
-    '/governance/oracle/proposals/:proposalId/approve/prepare',
-    idempotency,
-    (req, res, next) =>
-      prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-        const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-        const audit = validateGovernanceAuditInput(req.body);
-        const proposalId = validateProposalId(getPathParam(req.params.proposalId, 'proposalId'));
-        const proposal = await options.governanceReader.getOracleProposalState(proposalId);
-        if (!proposal) {
-          throw new GatewayError(404, 'NOT_FOUND', 'Oracle update proposal not found', {
-            proposalId,
-          });
-        }
-        if (proposal.executed || proposal.cancelled || proposal.expired) {
-          throw new GatewayError(
-            409,
-            'CONFLICT',
-            'Oracle update proposal is no longer approvable',
-            { proposalId },
-          );
-        }
-
-        if (await options.governanceReader.hasApprovedOracleProposal(proposalId, signerWallet)) {
-          throw new GatewayError(
-            409,
-            'CONFLICT',
-            'Caller has already approved this oracle update proposal',
-            { proposalId },
-          );
-        }
-
-        return options.mutationService.prepareAction({
-          category: 'oracle_update',
-          contractMethod: 'approveOracleUpdate',
-          routePath: req.originalUrl || req.path,
-          audit,
-          principal,
-          signerWallet,
-          requestContext,
-          signerBinding,
-          idempotencyKey,
-          proposalId,
-          targetAddress: proposal.targetAddress,
-        });
-      }),
-  );
-
-  router.post(
-    '/governance/oracle/proposals/:proposalId/execute/prepare',
-    idempotency,
-    (req, res, next) =>
-      prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-        const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-        const audit = validateGovernanceAuditInput(req.body);
-        const proposalId = validateProposalId(getPathParam(req.params.proposalId, 'proposalId'));
-        const proposal = await options.governanceReader.getOracleProposalState(proposalId);
-        if (!proposal) {
-          throw new GatewayError(404, 'NOT_FOUND', 'Oracle update proposal not found', {
-            proposalId,
-          });
-        }
-        if (proposal.executed || proposal.cancelled || proposal.expired) {
-          throw new GatewayError(409, 'CONFLICT', 'Oracle update proposal is not executable', {
-            proposalId,
-          });
-        }
-
-        const status = await options.governanceReader.getGovernanceStatus();
-        if (proposal.approvalCount < status.governanceApprovalsRequired) {
-          throw new GatewayError(
-            409,
-            'CONFLICT',
-            'Oracle update proposal does not have enough approvals',
-            { proposalId },
-          );
-        }
-        if (proposal.etaSeconds > nowSeconds()) {
-          throw new GatewayError(
-            409,
-            'CONFLICT',
-            'Oracle update proposal timelock has not elapsed',
-            { proposalId },
-          );
-        }
-
-        return options.mutationService.prepareAction({
-          category: 'oracle_update',
-          contractMethod: 'executeOracleUpdate',
-          routePath: req.originalUrl || req.path,
-          audit,
-          principal,
-          signerWallet,
-          requestContext,
-          signerBinding,
-          idempotencyKey,
-          proposalId,
-          targetAddress: proposal.targetAddress,
-        });
-      }),
-  );
-
-  router.post(
-    '/governance/oracle/proposals/:proposalId/cancel-expired/prepare',
-    idempotency,
-    (req, res, next) =>
-      prepareAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-        const { principal, requestContext, idempotencyKey } = getMutationContext(req);
-        const audit = validateGovernanceAuditInput(req.body);
-        const proposalId = validateProposalId(getPathParam(req.params.proposalId, 'proposalId'));
-        const proposal = await options.governanceReader.getOracleProposalState(proposalId);
-        if (!proposal) {
-          throw new GatewayError(404, 'NOT_FOUND', 'Oracle update proposal not found', {
-            proposalId,
-          });
-        }
-        if (proposal.executed || proposal.cancelled || !proposal.expired) {
-          throw new GatewayError(
-            409,
-            'CONFLICT',
-            'Oracle update proposal is not cancellable as expired',
-            { proposalId },
-          );
-        }
-
-        return options.mutationService.prepareAction({
-          category: 'oracle_update',
-          contractMethod: 'cancelExpiredOracleUpdateProposal',
-          routePath: req.originalUrl || req.path,
-          audit,
-          principal,
-          signerWallet,
-          requestContext,
-          signerBinding,
-          idempotencyKey,
-          proposalId,
-          targetAddress: proposal.targetAddress,
-        });
-      }),
-  );
-
-  // ── Broadcast confirm endpoint ────────────────────────────────────────────
-  // Called by the dashboard after the admin wallet signs and broadcasts the tx.
-  // Accepts the txHash, transitions the action into the monitored broadcast
-  // lifecycle, and records the final signer evidence for backend reconciliation.
-
-  router.post('/governance/actions/:actionId/confirm', (req, res, next) =>
-    confirmAndRespond(req, res, next, options.config, async (signerWallet, signerBinding) => {
-      if (!req.gatewayPrincipal) {
-        throw new GatewayError(401, 'AUTH_REQUIRED', 'Authentication is required');
-      }
-      if (!req.requestContext) {
-        throw new GatewayError(500, 'INTERNAL_ERROR', 'Request context was not initialized');
-      }
-
-      const actionId = getPathParam(req.params.actionId, 'actionId');
-      if (!actionId) {
-        throw new GatewayError(400, 'VALIDATION_ERROR', 'Path parameter actionId is required');
-      }
-
-      const body = req.body as Record<string, unknown>;
-      if (typeof body?.txHash !== 'string' || !body.txHash) {
-        throw new GatewayError(400, 'VALIDATION_ERROR', 'txHash is required');
-      }
-
-      return options.mutationService.confirmBroadcast({
-        actionId,
-        txHash: body.txHash,
-        signerWallet,
-        principal: req.gatewayPrincipal,
-        signerBinding,
-        requestContext: req.requestContext,
-      });
-    }),
-  );
+  registerGovernanceDirectSignRoutes(router, idempotency, options);
 
   return router;
 }
